code_stego/pyproject.toml
[tool.poetry]
name = "code-stego"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
numpy = "^1.25.2"
Pillow = "^10.0.0"
Pygments = "^2.16.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
